requirements.txt
# Slack
slack-bolt>=1.18.0
slack-sdk>=3.21.0

# AI
google-genai>=1.0.0

# Browser automation
playwright>=1.40.0

# Google Calendar
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# System tray
pystray>=0.19.0
Pillow>=10.0.0

# Configuration
PyYAML>=6.0.0

# Utilities
python-dateutil>=2.8.0

# Packaging (development)
pyinstaller>=6.0.0
